Add indexes identified in constraints to Meta field in inspection
When inspecting tables, the process creates a constraints dict
that identifies indexes. If we have any such indexes, let's add
them to `Meta.indexes` using the base `Index` class.

At this point, it does not support identifying the index type and
substituting in the proper index class, and I don't believe we have
the information at this time to select the right class anyway.

Ref: https://code.djangoproject.com/ticket/27060